+++ server/internal/model/entity/knowledge_documents.go
@@ -10,7 +10,7 @@ import (
 
 // KnowledgeDocuments is the golang structure for table knowledge_documents.
 type KnowledgeDocuments struct {
-	Id                int         `json:"id"                orm:"id"                  description:""` //
+	Id                int64       `json:"id"                orm:"id"                  description:""` //
 	KnowledgeBaseName string      `json:"knowledgeBaseName" orm:"knowledge_base_name" description:""` //
 	FileName          string      `json:"fileName"          orm:"file_name"           description:""` //
 	Status            int         `json:"status"            orm:"status"              description:""` //
